+++ Pinny Notes/Tools/_Interfaces/ITool.cs
@@ -4,5 +4,8 @@ namespace Pinny_Notes.Tools;
 
 public interface ITool
 {
+    bool IsEnabled { get; }
+    bool IsFavourite { get; }
+
     MenuItem GetMenuItem();
 }